# Fix `unit-no-unknown` false positives for the second and subsequent `image-set()` with `x` descriptor