--- a/lib/rules/unit-no-unknown/__tests__/index.js
+++ b/lib/rules/unit-no-unknown/__tests__/index.js
@@ -228,6 +228,18 @@ testRule({
 			description: 'ignore `x` unit in vendor-prefixed image-set',
 		},
 		{
+			code: "a { background-image: url('first.png'), image-set(url('second.png') 1x), image-set(url('third.png') 1x) }",
+			description: 'ignore `x` unit in the second and subsequent image-set',
+		},
+		{
+			code: "a { background-image: url('first.png'), -webkit-image-set(url('second.png') 1x), -webkit-image-set(url('third.png') 1x) }",
+			description: 'ignore `x` unit in the second and subsequent vendor-prefixed image-set',
+		},
+		{
+			code: "a { background-image: image-set(url('first.png') calc(1x * 1), url('second.png') calc(1x + 0.5x)); }",
+			description: 'ignore `x` unit in image-set with calc',
+		},
+		{
 			code: '@media (resolution: 2x) {}',
 			description: 'ignore `x` unit in @media with `resolution`',
 		},
@@ -443,9 +455,9 @@ testRule({
 			message: messages.rejected('x'),
 			description: '`x` rejected with inappropriate property',
 			line: 1,
-			column: 47,
+			column: 60,
 			endLine: 1,
-			endColumn: 48,
+			endColumn: 61,
 		},
 		{
 			code: "a { background: image-set('img1x.png' 1x, 'img2x.png' 2x) left 20x / 15% 60% repeat-x; }",