--- a/lib/rules/unit-no-unknown/index.js
+++ b/lib/rules/unit-no-unknown/index.js
@@ -81,7 +81,7 @@ const rule = (primary, secondaryOptions) => {
 		 * @param {import('@csstools/css-parser-algorithms').ComponentValue} componentValue
 		 * @param {{ ignored: boolean, allowX: boolean }} state
 		 */
-		const walker = (node, getIndex, componentValue, state) => {
+		const check = (node, getIndex, componentValue, state) => {
 			if (isFunctionNode(componentValue)) {
 				const name = componentValue.getName();
 
@@ -157,7 +157,7 @@ const rule = (primary, secondaryOptions) => {
 						}
 					}
 
-					walker(atRule, atRuleParamIndex, entry.node, entry.state ?? state);
+					check(atRule, atRuleParamIndex, entry.node, entry.state ?? state);
 				}, state);
 			});
 		});
@@ -183,7 +183,7 @@ const rule = (primary, secondaryOptions) => {
 						allowX: isImageResolutionProp,
 					};
 
-					walker(decl, declarationValueIndex, componentValue, state);
+					check(decl, declarationValueIndex, componentValue, state);
 
 					if (!isFunctionNode(componentValue) && !isSimpleBlockNode(componentValue)) {
 						return;
@@ -192,7 +192,7 @@ const rule = (primary, secondaryOptions) => {
 					componentValue.walk((entry) => {
 						if (entry.state?.ignored) return;
 
-						walker(decl, declarationValueIndex, entry.node, entry.state ?? state);
+						check(decl, declarationValueIndex, entry.node, entry.state ?? state);
 					}, state);
 				});
 		});