--- a/lib/rules/unit-no-unknown/index.js
+++ b/lib/rules/unit-no-unknown/index.js
@@ -78,7 +78,7 @@ const rule = (primary, secondaryOptions) => {
 		 * @template {import('postcss').AtRule | import('postcss').Declaration} T
 		 * @param {T} node
 		 * @param {(node: T) => number} getIndex
-		 * @param {import('@csstools/css-parser-algorithms').ComponentValue} componentValue
+		 * @param {import('@csstools/css-parser-algorithms').FunctionNode | import('@csstools/css-parser-algorithms').TokenNode} componentValue
 		 * @param {{ ignored: boolean, allowX: boolean }} state
 		 */
 		const check = (node, getIndex, componentValue, state) => {
@@ -102,9 +102,9 @@ const rule = (primary, secondaryOptions) => {
 
 					return;
 				}
-			}
 
-			if (!isTokenNode(componentValue)) return;
+				return;
+			}
 
 			const [tokenType, , , endIndex, tokenValue] = componentValue.value;
 
@@ -149,6 +149,13 @@ const rule = (primary, secondaryOptions) => {
 				mediaQuery.walk((entry) => {
 					if (entry.state?.ignored) return;
 
+					if (!entry.state) {
+						entry.state = {
+							ignored: false,
+							allowX: false,
+						};
+					}
+
 					if (isMediaFeature(entry.node)) {
 						const name = entry.node.getName();
 
@@ -157,7 +164,9 @@ const rule = (primary, secondaryOptions) => {
 						}
 					}
 
-					check(atRule, atRuleParamIndex, entry.node, entry.state ?? state);
+					if (isFunctionNode(entry.node) || isTokenNode(entry.node)) {
+						check(atRule, atRuleParamIndex, entry.node, entry.state ?? state);
+					}
 				}, state);
 			});
 		});
@@ -183,7 +192,9 @@ const rule = (primary, secondaryOptions) => {
 						allowX: isImageResolutionProp,
 					};
 
-					check(decl, declarationValueIndex, componentValue, state);
+					if (isFunctionNode(componentValue) || isTokenNode(componentValue)) {
+						check(decl, declarationValueIndex, componentValue, state);
+					}
 
 					if (!isFunctionNode(componentValue) && !isSimpleBlockNode(componentValue)) {
 						return;
@@ -192,7 +203,9 @@ const rule = (primary, secondaryOptions) => {
 					componentValue.walk((entry) => {
 						if (entry.state?.ignored) return;
 
-						check(decl, declarationValueIndex, entry.node, entry.state ?? state);
+						if (isFunctionNode(entry.node) || isTokenNode(entry.node)) {
+							check(decl, declarationValueIndex, entry.node, entry.state ?? state);
+						}
 					}, state);
 				});
 		});