--- a/lib/rules/unit-no-unknown/index.js
+++ b/lib/rules/unit-no-unknown/index.js
@@ -149,17 +149,10 @@ const rule = (primary, secondaryOptions) => {
 				mediaQuery.walk((entry) => {
 					if (entry.state?.ignored) return;
 
-					if (!entry.state) {
-						entry.state = {
-							ignored: false,
-							allowX: false,
-						};
-					}
-
 					if (isMediaFeature(entry.node)) {
 						const name = entry.node.getName();
 
-						if (RESOLUTION_MEDIA_FEATURE_NAME.test(name)) {
+						if (entry.state && RESOLUTION_MEDIA_FEATURE_NAME.test(name)) {
 							entry.state.allowX = true;
 						}
 					}