--- a/lib/rules/unit-no-unknown/index.js
+++ b/lib/rules/unit-no-unknown/index.js
@@ -147,6 +147,7 @@ const rule = (primary, secondaryOptions) => {
 				};
 
 				mediaQuery.walk((entry) => {
+					if (!entry.state) return;
 					if (entry.state?.ignored) return;
 
 					if (isMediaFeature(entry.node)) {