--- a/lib/rules/unit-no-unknown/index.js
+++ b/lib/rules/unit-no-unknown/index.js
@@ -148,18 +148,19 @@ const rule = (primary, secondaryOptions) => {
 
 				mediaQuery.walk((entry) => {
 					if (!entry.state) return;
-					if (entry.state?.ignored) return;
+
+					if (entry.state.ignored) return;
 
 					if (isMediaFeature(entry.node)) {
 						const name = entry.node.getName();
 
-						if (entry.state && RESOLUTION_MEDIA_FEATURE_NAME.test(name)) {
+						if (RESOLUTION_MEDIA_FEATURE_NAME.test(name)) {
 							entry.state.allowX = true;
 						}
 					}
 
 					if (isFunctionNode(entry.node) || isTokenNode(entry.node)) {
-						check(atRule, atRuleParamIndex, entry.node, entry.state ?? state);
+						check(atRule, atRuleParamIndex, entry.node, entry.state);
 					}
 				}, state);
 			});
@@ -195,10 +196,12 @@ const rule = (primary, secondaryOptions) => {
 					}
 
 					componentValue.walk((entry) => {
-						if (entry.state?.ignored) return;
+						if (!entry.state) return;
+
+						if (entry.state.ignored) return;
 
 						if (isFunctionNode(entry.node) || isTokenNode(entry.node)) {
-							check(decl, declarationValueIndex, entry.node, entry.state ?? state);
+							check(decl, declarationValueIndex, entry.node, entry.state);
 						}
 					}, state);
 				});