--- a/lib/rules/unit-no-unknown/index.js
+++ b/lib/rules/unit-no-unknown/index.js
@@ -157,9 +157,7 @@ const rule = (primary, secondaryOptions) => {
 						if (RESOLUTION_MEDIA_FEATURE_NAME.test(name)) {
 							entry.state.allowX = true;
 						}
-					}
-
-					if (isFunctionNode(entry.node) || isTokenNode(entry.node)) {
+					} else if (isFunctionNode(entry.node) || isTokenNode(entry.node)) {
 						check(atRule, atRuleParamIndex, entry.node, entry.state);
 					}
 				}, state);