--- a/package.json
+++ b/package.json
@@ -162,7 +162,7 @@
     "@stylelint/remark-preset": "^4.0.0",
     "@types/balanced-match": "^1.0.2",
     "@types/css-tree": "^2.3.1",
-    "@types/debug": "^4.1.7",
+    "@types/debug": "^4.1.8",
     "@types/file-entry-cache": "^5.0.2",
     "@types/global-modules": "^2.0.0",
     "@types/globjoin": "^0.1.0",