<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See : https://github.com/stylelint/stylelint/issues/6848
Supersedes : https://github.com/stylelint/stylelint/pull/6726

> Is there anything in the PR that needs further explanation?

- ~~this will fail because it uses unreleased features of the media query parser~~
- also happens to resolve #6726


before :

```
npm run benchmark-rule unit-no-unknown true
Warnings: 0
Mean: 143.3089323125 ms
Deviation: 10.593647978441565 ms
```

after : 

```
npm run benchmark-rule unit-no-unknown true
Warnings: 0
Mean: 109.86287500000002 ms
Deviation: 9.320724134887374 ms
```