@romainmenke 2023-06-01
@ybiquitous I intended to just change the parser here, but this change also had the side effect of fixing an issue.

I propose changing the title and adding the correct changelog for that:

```md
Fix `unit-no-unknown` false positives for the second and subsequent `image-set()` with `x` descriptor
```

I would also add `chalkygames123` as a co-author as they wrote the tests that I used to verify the fix in https://github.com/stylelint/stylelint/pull/6726