@ybiquitous 2023-06-02
Yes, that's a good idea on https://github.com/stylelint/stylelint/pull/6879#issuecomment-1572262355 

Please update them!