@ScorpAL 2023-06-05
HI Guys.
Something wrong after update from 15.6.2 to 15.6.3 

```css
.test {
    margin-bottom: 20zrr;
}
```
With **stylelint 15.6.2**
CLI and VSCode reported:
`Unexpected unknown unit "zrr"  unit-no-unknown`

With **stylelint 15.6.3**
CLI:
no errors reported
VSCode:
`Unexpected unknown unit "zrr" (unit-no-unknown)Stylelint[unit-no-unknown](https://stylelint.io/user-guide/rules/unit-no-unknown)`